Split Language Enum from Analyzer file
