+++ tests/integration_test.rs
@@ -3,8 +3,9 @@ mod integration_test {
     mod rust_test;
     mod python_test;
 
-    use balpan::analyzer::{Analyzer,Language, Traversable};
+    use balpan::analyzer::{Analyzer, Traversable};
     use balpan::grammar::{fetch_grammars, build_grammars};
+    use balpan::language::Language;
 
     pub fn assert_analyzed_source_code(source_code: &str, expected: &str, language: &str) {
         fetch_grammars().unwrap();
